Address Claude findings
