+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "@omega-edit/core",
   "version": "2.0.0",
-  "private": "true",
+  "private": true,
   "description": "OmegaEdit Client and Server",
   "publisher": "CTC-OSS",
   "repository": {
@@ -14,7 +14,7 @@
   "author": "CTC-OSS",
   "license": "Apache-2.0",
   "engines": {
-    "node": "^24.0.0"
+    "node": ">=22"
   },
   "workspaces": [
     "packages/*"
@@ -28,7 +28,7 @@
   },
   "devDependencies": {
     "@biomejs/biome": "^2.5.1",
-    "@types/node": "^26.0.1",
+    "@types/node": "^22.20.0",
     "copy-webpack-plugin": "^14.0.0",
     "run-script-os": "^1.1.6",
     "ts-loader": "^9.6.2",
@@ -41,7 +41,7 @@
   },
   "resolutions": {
     "@types/eslint-scope": "8.4.0",
-    "@types/node": "26.0.1",
+    "@types/node": "22.20.0",
     "copy-webpack-plugin/**/ajv": "8.20.0",
     "copy-webpack-plugin/**/picomatch": "4.0.4",
     "copy-webpack-plugin/serialize-javascript": "7.0.6",
