+++ packages/ai/package.json
@@ -32,7 +32,7 @@
   "author": "CTC-OSS",
   "license": "Apache-2.0",
   "engines": {
-    "node": "^24.0.0"
+    "node": ">=22"
   },
   "scripts": {
     "clean": "rimraf dist",
@@ -50,7 +50,7 @@
     "@omega-edit/client": "2.0.0"
   },
   "devDependencies": {
-    "@types/node": "^26.0.0",
+    "@types/node": "^22.20.0",
     "rimraf": "^6.1.3",
     "vitest": "^4.1.9"
   }
