+++ packages/server/package.json
@@ -18,7 +18,7 @@
   "author": "CTC-OSS",
   "license": "Apache-2.0",
   "engines": {
-    "node": "^24.0.0"
+    "node": ">=22"
   },
   "scripts": {
     "test": "echo 'C++ server tests run via yarn workspace @omega-edit/client test'",
